Create README.md
I used to collect lots of interesting and useful links and resources to share with customers, colleagues and partners on all kinds of topics but maintaining the list was a PITA. I've created a number of "markdown" format files on some of the popular topic areas. I'll add my thoughts on each of the subjects - What is it? How does it work? What are the alternatives? And then follow that with a bunch of links.